+++ .eslintrc.json
@@ -23,6 +23,7 @@
 		],
 		"global-require": "off",
 		"import/prefer-default-export": "off",
+		"react/jsx-one-expression-per-line": "off",
 		"no-unused-expressions": ["error", { "allowTaggedTemplates": true }],
 		"import/extensions": [
 			"error",
