+++ src/screens/home/assets-panel/avatar-test.png
