+++ src/screens/home/finance-panel/FinancePanel.tsx
@@ -1,0 +1,16 @@
+import React, { useState } from 'react';
+import TabSelector from './tab-selector/TabSelector';
+import NetWorth from './net-worth/NetWorth';
+import Transactions from './transactions/Transactions';
+
+const FinancePanel = () => {
+	const [selectedTab, setSelectedTab] = useState('transactions');
+	return (
+		<>
+			<TabSelector selectedTab={selectedTab} onTabSelect={setSelectedTab} />
+			{selectedTab === 'net_worth' ? <NetWorth /> : <Transactions />}
+		</>
+	);
+};
+
+export default FinancePanel;
