+++ src/screens/home/styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		flex: 1,
+		backgroundColor: '#F2EAE1'
+	},
+	homeScroll: {
+		marginBottom: 80
+	}
+});
