--- src/screens/home/styles.tsx
@@ -1,225 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-	container: {
-		flex: 1,
-		backgroundColor: '#F2EAE1'
-	},
-	homeScroll: {
-		marginBottom: 80
-	},
-	row: {
-		flexDirection: 'row',
-		justifyContent: 'space-between'
-	},
-	alignContentRight: {
-		alignItems: 'flex-end'
-	},
-	fontSizeDefault: {
-		fontSize: 16
-	},
-	fontSizeSmall: {
-		color: 'rgba(78, 94, 111, 1)',
-		fontSize: 12
-	},
-	fontBold: {
-		fontFamily: 'Inter_800ExtraBold'
-	},
-	paddingContent: {
-		paddingTop: 24,
-		paddingLeft: 24,
-		paddingRight: 24,
-		paddingBottom: 16
-	},
-	card: {
-		backgroundColor: 'rgba(255, 255, 255, 0.8)',
-		borderRadius: 24,
-		overflow: 'hidden',
-		marginBottom: 16
-	},
-	cardTopContent: {
-		padding: 24,
-		flexDirection: 'row',
-		justifyContent: 'space-between',
-		alignItems: 'center'
-	},
-	cardLabel: {
-		color: 'rgba(78, 94, 111, 1)',
-		fontSize: 14,
-		marginBottom: 8
-	},
-	cardBalance: {
-		color: 'rgba(10, 32, 54, 1)',
-		fontSize: 36
-	},
-	avatar: {
-		width: 48,
-		height: 48,
-		borderColor: 'rgba(243, 247, 255, 1)',
-		borderRadius: 50,
-		borderWidth: 4
-	},
-	cardBottomContent: {
-		borderTopColor: 'rgba(242, 234, 225, 1)',
-		borderTopWidth: 1,
-		backgroundColor: '#ffffff',
-		paddingLeft: 24,
-		paddingRight: 24,
-		flexDirection: 'row',
-		justifyContent: 'space-around'
-	},
-	cardActionButton: {
-		flexDirection: 'row',
-		alignItems: 'center',
-		paddingTop: 16,
-		paddingBottom: 16
-	},
-	cardButtonIcon: {
-		paddingRight: 8
-	},
-	cardDivisor: {
-		width: '50%',
-		flexDirection: 'row',
-		alignItems: 'center',
-		borderRightColor: 'rgba(242, 234, 225, 1)',
-		borderRightWidth: 1
-	},
-	scrollviewHorizontal: {
-		paddingLeft: 24,
-		marginBottom: 32,
-		paddingBottom: 8
-	},
-	scrollviewHorizontalContent: {
-		flexDirection: 'row',
-		paddingRight: 32
-	},
-	cardScroll: {
-		borderRadius: 16,
-		paddingLeft: 16,
-		paddingRight: 16,
-		marginRight: 16
-	},
-	appBar: {
-		backgroundColor: 'transparent',
-		elevation: 0
-	},
-	appBarContent: {
-		paddingLeft: 24,
-		paddingRight: 24,
-		flexDirection: 'row',
-		alignItems: 'center'
-	},
-	appBarIcons: {
-		flexDirection: 'row'
-	},
-	appBarIcon: {
-		paddingLeft: 24,
-		justifyContent: 'center'
-	},
-	appBarUserName: {
-		fontSize: 24,
-		color: '#0A2138',
-		fontFamily: 'Inter_800ExtraBold'
-	},
-	transactionDayRow: {
-		paddingTop: 24,
-		paddingBottom: 24,
-		borderBottomWidth: 1,
-		borderBottomColor: '#F2EAE1',
-		marginBottom: 24
-	},
-	transactionItem: {
-		marginBottom: 32,
-		flexDirection: 'row',
-		justifyContent: 'space-between'
-	},
-	transactionDateLabel: {
-		color: 'rgba(78, 94, 111, 1)',
-		fontFamily: 'Inter_800ExtraBold'
-	},
-	transationalIcon: {
-		width: 32,
-		height: 32,
-		marginRight: 8
-	},
-	tabsTransactions: {
-		paddingTop: 24,
-		paddingLeft: 24,
-		paddingRight: 24,
-		paddingBottom: 16,
-		backgroundColor: '#fff',
-		borderTopRightRadius: 24,
-		marginTop: -1
-	},
-	tabsNetWorth: {
-		paddingTop: 24,
-		paddingLeft: 24,
-		paddingRight: 24,
-		paddingBottom: 16,
-		backgroundColor: '#fff',
-		borderTopLeftRadius: 24,
-		marginTop: -1
-	},
-	netWorthItem: {
-		flexDirection: 'row',
-		justifyContent: 'space-between',
-		alignItems: 'center',
-		marginBottom: 32
-	},
-	netWorthIcon: {
-		padding: 16,
-		borderRadius: 16,
-		backgroundColor: '#F5F6F9'
-	},
-	netWorthIconAlert: {
-		padding: 16,
-		borderRadius: 16,
-		backgroundColor: '#F6E7E5'
-	},
-	netWorthItemText: {
-		flex: 1,
-		marginLeft: 16
-	},
-	arrowPadding: {
-		paddingLeft: 4,
-		marginTop: 2
-	},
-	roundInside: {
-		position: 'relative',
-		marginTop: 22
-	},
-	currentValueCard: {
-		paddingTop: 16,
-		paddingLeft: 24,
-		paddingRight: 24,
-		paddingBottom: 16,
-		backgroundColor: '#F5F6F9',
-		borderRadius: 24,
-		marginBottom: 40,
-		marginTop: 16
-	},
-	tabActive: {
-		flex: 1,
-		padding: 16,
-		alignItems: 'center',
-		borderTopLeftRadius: 16,
-		borderTopRightRadius: 16,
-		backgroundColor: '#fff',
-		fontFamily: 'Inter_800ExtraBold'
-	},
-	tabTitleActive: {
-		fontSize: 16,
-		fontFamily: 'Inter_800ExtraBold'
-	},
-	tabTitleInactive: {
-		fontSize: 16,
-		color: '#4E5E6F'
-	},
-	tabInactive: {
-		flex: 1,
-		padding: 16,
-		alignItems: 'center',
-		borderTopLeftRadius: 16,
-		borderTopRightRadius: 16
-	}
-});
